refactor(app): extract PrCaches into app/prs.rs
Struct and methods (detail_for, current, is_loading, set,
invalidate) move verbatim; re-exported from app. Part of #270.
